Update AlbFatherQW
This is an example of a simple translation I made. I tried not to change anything in the text that didn't need changing but there are words like Barlovskog and Zvezdushki that have no english analogue being names of areas or towns.  Also, some of the Russian names (surnames for example) are spelled differently for a male and a female (thus Albina and her Father having 'different' last names). Check it out if you like it I can go on with translating more.